+++ cairosvgutils.hpp
@@ -9,7 +9,7 @@
 #include "nanosvg.h"
 
 void draw_svg_image(cairo_t*, struct NSVGimage*,
-  double = 0, double = 0, int = 0, int = 0) noexcept;
+  double = 0, double = 0, double = 0, double = 0) noexcept;
 void draw_svg_image(Fl_Image*, struct NSVGimage*,
   double = 0, double = 0) noexcept;
